backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
pymysql
passlib
python-dotenv
bcrypt
smtplib
python-decouple
python-jose[cryptography]